Renamed several core files.
